Refactored ScannerPage & LoginPage
